Refactor directory management
- components > component dir
