+++ client/src/components/Cards/Cards.css
@@ -1,0 +1,8 @@
+.Cards {
+  margin: 5px;
+  height: 50%;
+  width: 10%;
+  border-style: solid;
+  border-width: 1px;
+  border-color: rgb(245, 225, 0);
+}
