+++ client/src/components/GameInfo/GameInfo.css
